Phase 79 (HS-79-03): routes/primitives.py becomes the primitives package — done
1,294 lines, 41 routes, one builder -> seven family routers (notes 97,
agents 206, profiles 103, kbs 97, chains 214, workflows 333,
directories 167) + _shared 159 (the run frame/persist tail stays ONE
function — agents, chains, and workflows import it, never fork it),
composed under the unchanged build_primitives_router. The package root
re-exports the pinned wire vocabulary (CANONICAL_SOURCE_TYPES,
canonical_source_type) — the vocabulary test caught the missing
re-export, then passed unmodified.

Verbatim by the programmatic check (zero code-body lines differ; 49
lazy imports one level deeper + the one sibling workflow_graph import,
caught by the workflow-run tests). Manifest diff module-fields-only.
Tests unmodified: unit 2407, integration 685.
